+++ tests/testthat/test_VonBertalanffy.R
@@ -8,7 +8,7 @@ context("Von Bertalanffy")
 
 test_that("vbFuns errors and warnings",{
   ## wrong types
-  expect_that(vbFuns(type="Derek"),throws_error())
+  expect_error(vbFuns(type="Derek"))
 })
   
 test_that("vbStarts errors and warnings",{
@@ -16,48 +16,33 @@ test_that("vbStarts errors and warnings",{
   if (require(fishmethods)) {
     data(Kimura)
     ## wrong types
-    expect_that(vbStarts(length~age,data=Kimura,type="Derek"),throws_error())
-    expect_that(vbStarts(length~age,data=Kimura,type="Francis",methEV="Derek"),throws_error())
-    expect_that(vbStarts(length~age,data=Kimura,type="Schnute",methEV="Derek"),throws_error())
-    expect_that(vbStarts(length~age,data=Kimura,type="typical",meth0="Derek"),throws_error())
-    expect_that(vbStarts(length~age,data=Kimura,type="original",meth0="Derek"),throws_error())
+    expect_error(vbStarts(length~age,data=Kimura,type="Derek"))
+    expect_error(vbStarts(length~age,data=Kimura,type="Francis",methEV="Derek"))
+    expect_error(vbStarts(length~age,data=Kimura,type="Schnute",methEV="Derek"))
+    expect_error(vbStarts(length~age,data=Kimura,type="typical",meth0="Derek"))
+    expect_error(vbStarts(length~age,data=Kimura,type="original",meth0="Derek"))
     ## Two variables on LHS
-    expect_that(vbStarts(length+age~age,data=Kimura,type="typical"),throws_error())
+    expect_error(vbStarts(length+age~age,data=Kimura,type="typical"))
     ## Two variables on RHS
-    expect_that(vbStarts(length~age+sex,data=Kimura,type="typical"),throws_error())
+    expect_error(vbStarts(length~age+sex,data=Kimura,type="typical"))
     ## LHS is a factor
-    expect_that(vbStarts(sex~age,data=Kimura,type="typical"),throws_error())
+    expect_error(vbStarts(sex~age,data=Kimura,type="typical"))
     ## RHS is a factor
-    expect_that(vbStarts(length~sex,data=Kimura,type="typical"),throws_error())
+    expect_error(vbStarts(length~sex,data=Kimura,type="typical"))
     ## not two ages2use given
-    expect_that(vbStarts(length~age,data=Kimura,type="Francis",ages2use=2),throws_error())
-    expect_that(vbStarts(length~age,data=Kimura,type="Francis",ages2use=c(2,5,10)),throws_error())
-    expect_that(vbStarts(length~age,data=Kimura,type="Schnute",ages2use=2),throws_error())
-    expect_that(vbStarts(length~age,data=Kimura,type="Schnute",ages2use=c(2,5,10)),throws_error())
+    expect_error(vbStarts(length~age,data=Kimura,type="Francis",ages2use=2))
+    expect_error(vbStarts(length~age,data=Kimura,type="Francis",ages2use=c(2,5,10)))
+    expect_error(vbStarts(length~age,data=Kimura,type="Schnute",ages2use=2))
+    expect_error(vbStarts(length~age,data=Kimura,type="Schnute",ages2use=c(2,5,10)))
     ## ages2use in wrong order
-    expect_that(vbStarts(length~age,data=Kimura,type="Francis",ages2use=c(10,2)),gives_warning())
-    expect_that(vbStarts(length~age,data=Kimura,type="Schnute",ages2use=c(10,2)),gives_warning())
+    expect_warning(vbStarts(length~age,data=Kimura,type="Francis",ages2use=c(10,2)))
+    expect_warning(vbStarts(length~age,data=Kimura,type="Schnute",ages2use=c(10,2)))
   }
   ## gives warning about a poor estimate for K and Linf
   if (require(FSAdata)) {
     data(SpottedSucker1)
     sv <- list(Linf=max(SpottedSucker1$tl),K=0.3,t0=0)
-    expect_that(vbStarts(tl~age,data=SpottedSucker1,type="typical"),gives_warning())
-  }
-})
-
-test_that("growthModelSim errors and warnings",{
-  ## Get some data for the following attempts
-  if (require(fishmethods)) {
-    data(Kimura)
-    ## Two variables on LHS
-    expect_that(growthModelSim(length+age~age,data=Kimura),throws_error())
-    ## Two variables on RHS
-    expect_that(growthModelSim(length~age+sex,data=Kimura),throws_error())
-    ## LHS is a factor
-    expect_that(growthModelSim(sex~age,data=Kimura),throws_error())
-    ## RHS is a factor
-    expect_that(growthModelSim(length~sex,data=Kimura),throws_error())
+    expect_warning(vbStarts(tl~age,data=SpottedSucker1,type="typical"))
   }
 })
 
@@ -66,17 +51,17 @@ test_that("walfordPlot and chapmanPlot errors and warnings",{
   if (require(fishmethods)) {
     data(Kimura)
     ## Two variables on LHS
-    expect_that(walfordPlot(length+age~age,data=Kimura),throws_error())
-    expect_that(chapmanPlot(length+age~age,data=Kimura),throws_error())
+    expect_error(walfordPlot(length+age~age,data=Kimura))
+    expect_error(chapmanPlot(length+age~age,data=Kimura))
     ## Two variables on RHS
-    expect_that(walfordPlot(length~age+sex,data=Kimura),throws_error())
-    expect_that(chapmanPlot(length~age+sex,data=Kimura),throws_error())
+    expect_error(walfordPlot(length~age+sex,data=Kimura))
+    expect_error(chapmanPlot(length~age+sex,data=Kimura))
     ## LHS is a factor
-    expect_that(walfordPlot(sex~age,data=Kimura),throws_error())
-    expect_that(chapmanPlot(sex~age,data=Kimura),throws_error())
+    expect_error(walfordPlot(sex~age,data=Kimura))
+    expect_error(chapmanPlot(sex~age,data=Kimura))
     ## RHS is a factor
-    expect_that(walfordPlot(length~sex,data=Kimura),throws_error())
-    expect_that(chapmanPlot(length~sex,data=Kimura),throws_error())
+    expect_error(walfordPlot(length~sex,data=Kimura))
+    expect_error(chapmanPlot(length~sex,data=Kimura))
   }
 })