+++ NEWS.md
@@ -1,9 +1,10 @@
 # FSA 0.5.3 ongoing
+* `growthModelSim()`: Deleted.  The simulation functionality was moved to the `FSAsims` package.  The functionality related to finding starting values for the von Bertalanffy modesl was moved to `vbStarts()`.
 * `srFuns()`: Modified.  A complete rebuild to make similar to `vbFuns()`.  Added `simple=`.  Added `type='Shepherd'` for the Shepherd (1982) three parameter model and `type='SailaLorda'` for the "Saila-Lorda" three parameter model from Iles (1994).  Added tests for error messages.
 * `srModels()`: Modified.  A complete rebuild to make similar to `growthModels()`.  Added "Shepherd" and "Saila-Lorda" models.
 * `srSims()`: Deleted.  The simulation functionality was moved to the `FSAsims` package.  The functionality related to finding starting values was moved to `srStarts()`.
 * `srStarts()`: Modified.  A complete rebuild to streamline.  Removed default method (i.e., a formula must be used now).  Added "Shepherd" and "Saila-Lorda" models.  Modified plotting routine, including adding `col.mdl=`, `lwd.mdl=`, and `lty.mdl=`.  Moved the dynamic modeling aspects of `srSim()` into this function and is called with the new argument `dynamicPlot=TRUE`.  Also added `minmax.ratio=` and `delta.prop=` for use with the dynamic plots.
-* `vbStarts(): Modified.  A complete rebuild to streamline and fix some bugs that have not been found.
+* `vbStarts()`: Modified.  A complete rebuild to streamline and fix some bugs that had not been found.  Modified plotting routine, including adding `col.mdl=`, `lwd.mdl=`, and `lty.mdl=`.  Also added all of the von Bertalanffy parameterizations in `growthModelSim()` into this function and is called with the new argument `dynamicPlot=TRUE`.  Added dynamics plots for the "Francis" and "Schnute" parameterizations.
 
 # FSA 0.5.2 Mar15
 * `psdPlot()`: Modified.  Fixed bug related to `NA`s in `max.brks` variable.